+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix a style issue with the workspace footer
 - Prevent objectives used in selections from being deleted
 - Fix an issue where modals misbehaved sporadically
+- Move "Many Students Wonder" from activity styling to content styling
 
 ### Enhancements